DL-16059: Renamed files and package to prevent incorrect imports
