+++ it/test/utilities/IntegrationSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 HM Revenue & Customs
+ * Copyright 2025 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,22 @@
  * limitations under the License.
  */
 
-package util
+package utilities
+/*
+ * Copyright 2025 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 import org.apache.pekko.actor.ActorSystem
 import org.apache.pekko.stream.{ActorMaterializer, Materializer}
@@ -28,12 +43,12 @@ import play.api.inject.guice.GuiceApplicationBuilder
 import uk.gov.hmrc.http.HeaderCarrier
 
 trait IntegrationSpec
-    extends PlaySpec
+  extends PlaySpec
     with GuiceOneServerPerSuite
     with ScalaFutures
     with IntegrationPatience
     with Matchers
-    with WiremockHelper
+    with WiremockHelperIT
     with BeforeAndAfterEach
     with BeforeAndAfterAll {
