+++ it/test/utilities/WiremockHelperIT.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 HM Revenue & Customs
+ * Copyright 2025 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,23 @@
  * limitations under the License.
  */
 
-package util
+package utilities
+/*
+* Copyright 2025 HM Revenue & Customs
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
 
 import com.github.tomakehurst.wiremock.WireMockServer
 import com.github.tomakehurst.wiremock.client.WireMock
@@ -22,20 +38,26 @@ import com.github.tomakehurst.wiremock.client.WireMock._
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig
 import com.github.tomakehurst.wiremock.stubbing.StubMapping
 
-trait WiremockHelper extends FakeConfig {
+
+trait WiremockHelperIT extends FakeConfig {
+
 
   val url            = s"http://$wiremockHost:$wiremockPort"
   val wmConfig       = wireMockConfig().port(wiremockPort)
   val wireMockServer = new WireMockServer(wmConfig)
+
 
   def startWiremock(): Unit = {
     wireMockServer.start()
     WireMock.configureFor(wiremockHost, wiremockPort)
   }
 
+
   def stopWiremock(): Unit = wireMockServer.stop()
 
+
   def resetWiremock(): Unit = WireMock.reset()
+
 
   def stubGet(url: String, status: Integer, body: String): StubMapping = {
     removeStub(get(urlMatching(url)))
@@ -47,6 +69,7 @@ trait WiremockHelper extends FakeConfig {
     )
   }
 
+
   def stubPost(url: String, status: Integer, responseBody: String): StubMapping = {
     removeStub(post(urlMatching(url)))
     stubFor(
@@ -57,6 +80,7 @@ trait WiremockHelper extends FakeConfig {
     )
   }
 
+
   def stubPatch(url: String, status: Integer, responseBody: String): StubMapping =
     stubFor(
       patch(urlMatching(url))
@@ -64,6 +88,7 @@ trait WiremockHelper extends FakeConfig {
           aResponse().withStatus(status).withBody(responseBody)
         )
     )
+
 
   def stubPut(url: String, status: Integer, responseBody: String): StubMapping = {
     removeStub(put(urlMatching(url)))
@@ -75,11 +100,14 @@ trait WiremockHelper extends FakeConfig {
     )
   }
 
+
 }
 
+
 trait FakeConfig {
-  val wiremockPort = 80
+  val wiremockPort = 8083
   val wiremockHost = "localhost"
+
 
   def fakeConfig(additionalConfig: Map[String, String] = Map.empty): Map[String, String] = Map(
     "auditing.consumer.baseUri.host"                      -> s"$wiremockHost",
@@ -92,5 +120,6 @@ trait FakeConfig {
     "microservice.services.citizen-details.port"          -> s"$wiremockPort",
     "microservice.services.citizen-details.checkRequired" -> "true"
   ) ++ additionalConfig
+
 
 }
